Updated readme, used more brewisms

+++ scripts/make-grpc.sh
@@ -12,28 +12,42 @@ BUILD_DIR="${ROOT_DIR}/build/grpc"
 
 # Hop into a fresh build dir
 mkdir -p "${BUILD_DIR}"
-rm -rf "${BUILD_DIR}/"*
+rm -rf "${BUILD_DIR:?}/"*
 cd "${BUILD_DIR}"
 
 # Find openssl at a set of locations we expect
-OPENSSL_ROOT_DIR=""
-OPENSSL_DIRS=(
+export OPENSSL_ROOT_DIR=""
+OPENSSL_DIRS=()
+
+# Use the mongodbtoolchain if we have it
+OPENSSL_DIRS+=(
     "/opt/mongodbtoolchain/v2"
-    "/usr/local/opt/openssl@1.1"
-    "/usr/local/opt/openssl"
+)
+
+# Check brew just in case
+if command -V brew; then
+    OPENSSL_DIRS+=(
+        "$(brew --prefix openssl@1.1)"
+        "$(brew --prefix openssl)"
+    )
+fi
+
+# Otherwise, use the system dirs
+OPENSSL_DIRS+=(
     "/usr/local"
     "/usr"
 )
+
+# Search through our dirs in order to find our openssl
 for DIR in "${OPENSSL_DIRS[@]}"; do
     OPENSSL_EXE="${DIR}/bin/openssl"
     if [[ ! -f $OPENSSL_EXE ]]; then
         continue
     fi
 
-    OPENSSL_ROOT_DIR="-DOPENSSL_ROOT_DIR:PATH=${DIR}"
+    OPENSSL_ROOT_DIR="${DIR}"
     break
 done
 
-# Use the mongodbtoolchain openssl to build and install grpc
-cmake "${OPENSSL_ROOT_DIR}" "${CMAKE_ARGS[@]:-}" "${ROOT_DIR}"
+cmake "${CMAKE_ARGS[@]:-}" "${ROOT_DIR}"
 make project_grpc
